Add support for nested rpms.
If a rpm exists in the extracted package, recursively call the script
to do a comparison of the rpm.

It would be better if rpm wasn't a special case, but the script is
fragile due to global variables and this is the safest method for now.

Although typically an rpm won't contain another rpm, a gzipped file or
other type of package may.

Signed-off-by: Randy Witt <randy.e.witt@linux.intel.com>