rpm-check.sh: Error when one of the rpms is not found
The code for this was already there, but if "readlink -f" returned an
empty string, the test passes even though the file doesn't exist. So fix
the test to actually use a quoted string which will error.

The error message also had to be changed to actually print the actual
arguments rather than the output from readlink, since the error on an
empty result from readlink would just print "can't open ".

Signed-off-by: Randy Witt <randy.e.witt@linux.intel.com>