jar: always list files without date+size
so that results on systems with unzip and fastjar are the same

Without this patch, in the unzip case,
files of same size were randomly ordered (e.g. in log4j-mini)
and build-compare would report a diff where it would not with fastjar.

This is also nice, because it allows to drop some code.