add functions to touch host watchdog
The host watchdog waits for newlines in the buildlog output.
The build is considered stuck if no newline was seen for 3600 seconds.
Output without newline, like 'echo -n .', is not recognized as progress.

Add two helpers to touch the host watchdog if required.
Add custom 'echo' function to reset timer.

Signed-off-by: Olaf Hering <olaf@aepfle.de>