fix objdump parsing
The path to the ELF binary will be part of the output. Since it
contains the old/new path elements, the output would always differ.
The previous way to hide this difference was wrong, it would match
also the ASCII output of the hexdump.

By doing a chdir prior objdump, the path to the ELF binary will be
identical.

Fixes #52

Signed-off-by: Olaf Hering <olaf@aepfle.de>